requirements.txt
Flask
pdfminer.six
pytesseract
Pillow
python-docx
spacy
openai
gunicorn
camelot-py[cv]